Logging: update to latest einhard upstream
* fix indent color offset calculation

The sizeof operator previously just determined the size of a pointer -
which coincided with the size of the ANSI color code. But that sure
isn't portable. So I fixed the logic to use a sizeof("") expression.
Also added a multiline output test.

* restore the fill character and set it only once

ostream::fill persists. Therefore it must be set only once, and reset to
what it was before.

* support for systems that don't support thread_local

Mac OS apparently does not support thread-local storage. In this case
have the Output object use a local stringstream object.

* remove thread_local stringstream declaration

The declaration is not needed by any code in the header (or even by code
outside of einhard). Thus it's better to hide the symbol from others (in
an anonymous namespace the symbol cannot be accessed from outside at
all).

refs #853

Signed-off-by: Matthias Kretz <kretz@kde.org>