clean up inputfunctions.h
The unnamed namespace lead to working code, but wasn't conceptually the
right thing to use here. Instead most of the functions were meant to be
inlined. But the line_parser function is so large that it really should
be in its own source file (which I added as inputfunctions.cc).

Signed-off-by: Matthias Kretz <kretz@kde.org>