Actions/ParticleData: drop collision_time member
* The collision time should be fully handled via the
Action::time_of_execution_ value, which is used as the sort key for
actions.
* The collision_time member of ParticleData enabled an early detection
of actions that would act on the same particle but at a later time.

The problem with the old approach was that it was dependent on the order
of the particles: i.e. ScatterAction iterated over the particle pairs
and the pairs that came first would always lead to an Action object,
even if the collision_time was later than a pair that was checked in
some subsequent iteration. Once we go concurrent, there won't be any
well-defined order anymore and thus reproducibility cannot work for a
fixed random seed.

Furthermore, the Action finder step should keep the Particle map
immutable. Conceptionally, finding means looking at something, not
modifying. The collision_time member of ParticleData required the finder
to write to that value for each Action (candidate) it created.

Finally, wasn't this actually a buggy implementation? Consider that
ScatterActionsFinder creates the following Actions:
* Particle a and b -> ScatterAction at time 2
* Particle a and c -> ScatterAction at time 1
* Particle b and d -> no Action because collision time would be 3
When the Actions are executed:
* time 1: a and c scatter
* time 2: a and b don't scatter because a's momentum changed
* time 3: b and d should scatter but there's no Action for it ...

refs #912

Signed-off-by: Matthias Kretz <kretz@kde.org>