DecayAction: use a random decay time
* Ensure that DecayAction has a time_of_execution \in [0, dt[. They had
a hard-coded 0 before. We reuse the random number that was used to
determine whether the particle decays at all for the time of the decay.
* Drop 1-param ctor as it is not used (which typically equals not
tested). Also, the action should always have a time_of_execution,
therefore using the 1-param ctor overload is wrong.

With time_of_execution = 0 the actions were performed before any
ScatterActions and with arbitrary ordering of decays. This lead to
different decay channels being taken for otherwise equivalent runs (if
std::sort decided to sort differently).

refs #925, #928

Signed-off-by: Matthias Kretz <kretz@kde.org>