Grid: add a simple Grid class
* The Grid uses a (currently fixed) cell length of 2.5 fm to partition
the existing particles into cells. There is a template argument to
remind about the need to support periodic boundaries from the Grid code
as well. This is currently not implemented, though.
* new log area for "Grid"
* I had to modify the find_possible_actions signature again in order to
avoid full concatenation of all the neighboring cells to a single
ParticleList. This would lead to a lot of copying of ParticleData
objects. The current solution simply references all the cells (which are
ParticleList objects) to be considered as neighbors.
* Experiment::run_time_evolution now creates a Grid (Normal) object in
every time step. This grid is used to iterate over the search_list and
neighbors_list, which are then passed to the action finder
find_possible_actions functions.
* ScatterActionsFinder::find_possible_actions needed a smarter
implementation to loop over the expected particle pairings. This was
much easier to achieve via a lambda expression than building an iterator
class.
* The iterate_all_pairs helper function ensures that the order of
arguments to the closure are sorted by the Particle ID. This is only
necessary to make it easier to reproduce the same output as master (or
different Grid cell lengths).

Speedup of a Box run with:
* 1200 particles
* 10³ fm³
* 100 time steps
* 2.5 fm grid cell length
* broken periodic boundaries...

=> 29.5 s vs. 4.5 s = 6.5 ⨉ speedup

with 1.5 fm grid cell length:
=> 29.5 s vs. 1.5 s = 19.7 ⨉ speedup

refs #186, #738, #737

Signed-off-by: Matthias Kretz <kretz@kde.org>