DecayAction: one less new/delete in the common case
Most iterations of DecayActionsFinder determine that no decay is to be
performed. Thus the DecayAction that was created on the heap must be
deleted again. Instead the DecayAction can be created on the stack and
in the rare case that it is to be returned from the function it is moved
onto the heap.

Have Action implement the move constructor efficiently.

I actually expected the compiler to do this automatically, but at least
GCC 4.8.1 didn't. The move ctor for Action must move the std::vector
members and not copy them in order to get any performance benefit.

I disabled the copy ctor explicitly and documented that one should
either move or create a new action. Copying is not a good usage pattern
for Action classes.

Signed-off-by: Matthias Kretz <kretz@kde.org>