Grid: keep the number of cells reasonable
I tested with a 1000³ fm³ Box and SMASH with Grid was a lot slower than
without the Grid. Which is obvious once you realize that it created
400x400x400 = 64000000 cells. But the Box had only 600 particles. That's
just plain crazy.
So I calculate the cubic root of the number of particles to determine an
upper bound for the number of cells in one dimension. This required a
different implementation for calculating the indexes, which also should
be more efficient because I replaced three float divisions with
multiplications.

refs #186, #738

Signed-off-by: Matthias Kretz <kretz@kde.org>