Grid: fix grid construction for Collider
if min and max position are equal the number_of_cells in one direction
would be zero and thus no cells would be constructed. On filling the
cells there would be an out-of-bounds access right away. This fixes cell
construction for this cases by enforcing a minimum number of 1 cell per
direction

refs #186, #738

Signed-off-by: Matthias Kretz <kretz@kde.org>