Grid: don't crash for a dilute system
The cell iteration was emitting neighbor cells even though there were
none. At least for the single-cell case this is now tested to work
correctly.

This commit contains a rewrite of the Grid::iterate_cells function. The
previous one was wrong and unmaintainable spaghetti code. Now the code
uses several function calls (actually lambdas - in order to capture
local variables) to determine the necessary iterations in each direction
of space.

refs #186, #738

Signed-off-by: Matthias Kretz <kretz@kde.org>