Particles Map Iterator: fix member types
the member types of the iterator classes were simply inherited from the
map iterators. In that case they were pointing to pairs instead of the
ParticleData entry of the pair - which the adaptor returns. This was
actually checked by libc++ to be consistent - and thus usage of the
iterator failed to compile.

Signed-off-by: Matthias Kretz <kretz@kde.org>