tests: make mocking Configuration easier
Many tests need a Configuration object. Loading the YAML contents from
file is something that the Configuration unit test verifies. Here we
rely on Configuration doing the right thing and rather want to load from
YAML embedded into the test. Therefore I added an implicit constructor
that is only available to unit tests that can convert a const char * to
a Configuration object.

Signed-off-by: Matthias Kretz <kretz@kde.org>