PdgCode: major simplification and speedup
PdgCode::operator< is called from ParticleType::find, which is called
very very often (too much, I think). The ParticleType search uses
PdgCode::operator< to find the correct particle. Therefore this function
has to be as fast as possible. And comparing the dump_ uint32_t is just
that fast implementation that we enabled through the union-based
implementation.

refs #581

Signed-off-by: Matthias Kretz <kretz@kde.org>