ParticleType: new pointer-like class for ParticleType
* overload ParticleType::operator& to return ParticleTypePtr instead of
the normal ParticleType*. This inhibits (unless you know about
std::addressof) storing/passing pointers to ParticleType. We want to
avoid them because the ParticleType vector should be allowed to be
duplicated (e.g. for distribution on a cluster, offload to an
accelerator, or only faster NUMA performance).
* new ParticleTypePtr class that stores the index of the ParticleType
object inside the ParticleType vector. Storage type is a uint16_t, and
thus a factor 4 smaller than ParticleType* and a factor 2 smaller than
PdgCode (which never should be used as a ParticleType reference).
* ParticleTypePtr behaves like a ParticleType* by implementing operator*
and operator->.
* adjust several function parameters where a const ParticleType * was
expected to use const ParticleType & instead. These errors became
visible thanks to the operator& overload.

refs #1206

Signed-off-by: Matthias Kretz <kretz@kde.org>