BlattWeisskopf: optimized
* replaced serial calculations by parallel calculations: consider that a
multiplication can be issued every cycle but takes 4 cycles to complete.
This means we want to issue three independent multiplications before the
result of the first multiplication is used for the next operation.
Therefore, the multiplication sequence x*x*x*x*x*x*x*x requires 7 * 4
cycles to complete.
  const auto x2 = x * x;
  const auto x4 = x2 * x2;
  x4 * x4
on the other hand only needs 3 * 4 cycles. The compiler cannot transform
the first expression to the second because floating point operations
cannot be reordered. Therefore it has to evaluate left-to-right.
* replaced the constants in BlattWeisskopf with single-precision
literals instead of double-precision. The input and output type was
float already and there is no apparent reason to convert the whole thing
to double just for the calculation
* replaced sqrt(x2)*sqrt(x2) by x2. This is a straightforward
optimization to get rid of the unneeded sqrt.
* disabled the sanity check for correct L input from Release builds.
This simplifies the generated code as the compiler doesn't have to add
the code for stack unrolling.

refs #1238

Signed-off-by: Matthias Kretz <kretz@kde.org>