pCM: optimize
* instead of requiring double-precision, use a deduced template
parameter to allow both double- and single-precision usage. This
requires the constants to be cast explicitly to T.
* changed C-style variable declaration to const auto declarations at
first use.
* replaced x*x/(4*s) by x*x*(0.25/s) since this allows the CPU to
execute the division much earlier and in parallel to the calculation of
x (and x*x). The division is the most complex operation (ignoring sqrt)
here and really should not be the last in the dependency chain.

refs #1238

Signed-off-by: Matthias Kretz <kretz@kde.org>