cmake: add an option to disable the sanitizers
The sanitizers are great for code verification in general, except when
debugging in gdb and you want to understand the asm. In that case the
asan and usan code distracts from the actual functionality and makes the
asm really hard to follow. So now you can disable the sanitizers for
such debugging sessions.

Signed-off-by: Matthias Kretz <kretz@kde.org>