Particles: optimize data_ and iteration
Particles::data_ was a std::vector and thus didn't allow an easy
sentinel marker at the end of the used storage. This required to keep
one extra element allocated in the vector at the end, and required extra
effort to track the consistency. There's no alternative to a sentinel
entry if iterator compares are supposed to work with a single pointer
compare. So instead of blowing up the iterator I replaced the
std::vector with a "manually" allocated array of ParticleData. This
array is built in such a way that it's always larger than the number of
particles actually used, and all remaining entries can safely be
dereferenced from iterator::operator++ and inspected to determine
whether a hole in data_ must be skipped.

Identification of holes in data_ is now implemented via a new boolean in
ParticleData (which cost no extra storage requirement since it sits in
the padding after ParticleTypePtr). This boolean is true for holes in
the begin() end() range and false otherwise. It is very important that
the entry end() points to uses hole_ == false, since otherwise
iterator::operator++ would overshoot and miss the end() iterator.

* This commit fixes the iterator TODO I kept in the code.
* After this change a run that took 15 minutes now takes 10 minutes.
(using the default config.yaml with { Testparticles: 50, Nevents: 5 })
* Passes all unit tests.

I still have to write/update the comments.

refs #1462

Signed-off-by: Matthias Kretz <kretz@kde.org>