ParticleType: (re)move the exception classes.
The input parser for ParticleType can throw an exception which was
previously a member type or Particles. That's incorrect and therefore I
moved the exception class to ParticleType (and adjusted the includes and
the unit test accordingly). The second exception class (ParseError) was
not used at all and therefore I removed it.

Signed-off-by: Matthias Kretz <kretz@kde.org>