Grid: fix index_factor for PeriodicBoundaries
With the reduced cell count the index_factor must be corrected,
otherwise out-of-bounds accesses can happen. The index_factor is
therefore recalculated from the number_of_cells and length.
std::nextafter is used to ensure the index_factor is rounded down and
thus the corner case in the testcase does not fail anymore.

refs #1128

Signed-off-by: Matthias Kretz <kretz@kde.org>