Integrator: C++ify GSL integration
* with unit test
* no arguments and lambda captures work
* extra arguments passed via the void * current segfault
* Dropped the extra argument option to be passed via function arguments
to the callable. This was segfaulting before and had an awkward syntax
anyway. Whereas lambda captures solve the same problem and are much
nicer to use here.
* Documented the whole thing.
* Moved the extra parameters to gsl_integration_qag to class member
variables to have them easily configurable when needed. This makes the
Integrator objects slightly larger. But then again there won't likely be
many of these objects in memory. In terms of optimizing the
gsl_integration_qag function the constexpr parameters couldn't help
anyway because the GSL code is not inline.
* The Integrator function call operator returns a (value, error) pair
now - in case you ever care for the error. GSL always returns it - so
there is no extra cost by making it available in the interface here.
* Allow direct assignment to double. The Integrator::operator() return
type is implicitly convertible to double. This allows direct assignment
to a double variable/argument, discarding the error.

refs #1823

Signed-off-by: Matthias Kretz <kretz@kde.org>