Grid: drop the index_factor_ member
The index factor was/is only used at construction of the Grid. Therefore
it does not need to be a member. The corresponding make_index member
function can be implemented via a local lambda in build_cells. This
makes the Grid interface leaner and cleaner. At the same time the
missing adjustment of index_factor in the 1x1x1 case becomes obviously
irrelevant.

Signed-off-by: Matthias Kretz <kretz@kde.org>