DecayModes: decaymodes.cc owns the DecayTypes
The DecayBranch objects held a non-owning pointer to DecayType objects.
Where in fact the ownership was somehow managed in the decaymodes.cc
file in a cleanup function. The DecayBranch objects were using the
branches with const-ref semantics (i.e. can't every be nullptr and will
never change the pointer). Therefore, they are now const-refs to objects
owned by a std::vector allocated in DecayModes::load_decaymodes.

Adjusted the tests to compile without the clear functions. (Those
functions were not used other than to free the memory at the end of
main()).

refs #1662

Signed-off-by: Matthias Kretz <kretz@kde.org>