interpolation: simplify logic in find_index
By changing the implementation to return the index to the last value
that is strictly smaller (instead of smaller or equal), we can avoid a
comparison of floating point numbers. Another case was redundant.

For the purpose of interpolation the changed behavior does not make any
difference, because the interpolations are always continuous.