Rough draft of how an extra dilepton routine in experiment.cc might work
The experiment class now has two new members one for the dilepton output
and one for the dilepton finder, because you don't want them in the
standard list for outputs and finders. Both new members are treated
analog to the other finders/outputs. Had to change set_outputs to
now have two arguments.

The Finder now returns an (dilepton) action list and then a routine
similiar to perform_actions from experiment is used which finally
calls the dilepton output.

This revison is already deprecated, because the outputs are now
inistalized in experiment.cc, not anymore in smash.cc. I will
adopt that.

still to do:

* one central switch for dilepton
* extra perform_dilepton_actions for experiment.cc