Refinements of dilepton output section in experiment.cc
Integrated the new output initalization in experiment.cc. For
readability reason, I moved the actual output and the prepreations
in an extra experiment function (perform_dilepton_action). Also
changed the format of the weight in Oscar Collision Output (might
change that back).

Still to do: dilepton switch in config.yaml

this refs #2328