Dilepton block(s) in experiment finished
There are now two dilepton blocks in experiment one after the
other find_possible_actions and the other after find_final_actions.

They directly after finding the dilepton decays output them via
the write_dilepton_action function.

They are only performed, if a dilepton finder was initalized.

To switch dilepton finder and output on or off, one has a switch
in config.yaml in the Output section. It triggers the bool
variable dileptons_switch, which itself triggers the
initalization of dilepton finder and output.

In this particular branch find_final_actions does nothing, but I
already implemented this in the main dilepton branch.

this refs #2328