Dilepton dalitz decays: kinematics and finder
The main challenge for the Dalitz Decays is that you do not have
a nice formula for the decay width, instead you have a formula
for the differential decay width. This means you have to treat them
differently. Fortunately you do not have to integrate, because one
can use the differential width directly for the shinning weight.

DecayActionFinderDilepton now differentiates between two and
threebody decays and uses the differential width for threebody
decays. For the exact treatment of dilepton dalitz decays see
future wiki entry or documentation.

Also new is a special one_to_three function, which calculates
the kinematics for the dalitz decays. It works in two steps:
first it "decays" the parent particle into the non-lepton
final state particle and a virtual photon (has mass).
After that the virtual photon "decays" into the dilepton and
boosts back into the parent particle rest frame.

There is still work to be done:

* Implement the formula for the differential width of the omega
dilepton daltiz decay.
* Calculate the shining weight in the finder correctly.
* Look into the find_final_actions routine for dalitz decays.
* Make SMASH work with (virtual) photon decays.
* Minor cosmetic changes.

this refs #2121