Solution for non-decaying resonances
The problem is if you add dilepton decays, but you do not want
to perform them, you get e.g. rhos below the 2 pion threshold,
because their lightest decay channel is the dilepton one.

If you then block the actual dilepton decays, because you want
dilepton treatet via the shining method, the rhos cannot
decay, because you take the only possible channel away.

The solution is to let the rhos live until the find_final_actions
and then allow the dilepton decay.

For this the get_partial_widths_hadronic function is now used in
find_possible_actions and the normal get_partial_widths, which
includes all decay channel is only used in find_final actions.

This solves the last remaining issue for the shining method.

this refs #1617