Fixes for dilepton dalitz decays
* dilepton mass distribution has now correct range
* diff_width is 0 for parent particle below threshold
* omega diff_width now correct
* increased the precision of momentum components in oscar output,
because I got sometimes negative invariant dilepton masses due to
rounding errors (this increases the output file size by ~10%)

I still get the gaussian smearing warning and would get floating
point exceptions, because of very small negative values
in sqrt in abs(), which I correct manually to 0.0. The problem
seems to only occur, when the density output is turnend off, which
is strange. It needs further testing.

this refs #2121