Changed the design of the diff_width function
Because the diff_width is only of any use in the decaytype
ThreeBodyDecayDilepton it was a bit overkill to make it a
member function of the base class. So now it is only a member
of ThreeBodyDecayDilepton.
This requires a dynamic_cast in the DecayActionFinderDilepton,
but I think it is the better solution.

this refs #2121