Fixed load_decaymodes_no_decays test
The exception was problemtic because of the pi in decaymodes.txt,
which has only a dilepton decay channel for pi0, but nothing for
pi+ and pi-. This is now fixed by only throwing the Missing Decay
Exception for unstable particle. (Pi0 is stable concerning hadronic
decays.)

There is still something wrong with the decaymodes.txt, but, since
this is also the case on the master, which was merged into the
branch, this should be vanish with the current master.

this refs #2121