Replace SMASH_CONST with SMASH_PURE
Telling the compiler a function is const means it does not have side
effects and it does not use global memory. This means that the function
cannot take arguments by reference, leading to subtle bugs like #2588.

Using the compiler attribute `pure` instead is much safer, because it
allows the function to use global memory and passing by reference works
as expected.

Fixes and refs #2588.