Function to compute gradient lattice
Given the quantity A(x,y,z) on the lattice, newly created
function computes grad A(x,y,z). Computation is tested using
several functions with known gradients. The pupose is
to calculate potential gradient on the lattice, when potential is given.
That is why this commit belongs to this particular issue.

This refs #1628

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>