Optimize collision_time
Function "collision_time" is a major bottleneck of SMASH, it takes more than 80%
of time for elastic box setup. The bottleneck of the function appeared to be
calculation of the particle velocities.  p1/E1 and p2/E2 is alredy six
divisions. After rearranging p1/E1 - p2/E2 to (p1*E2 - p2*E1)/(E1*E2) and
performing the division only once in the end I got 3 times speed-up.
Collision time function was inlined to let comipler optimize it, which results
in a small speed-up (from 132.8 to 130.9 s for an elastic box with pions test
case due to inlining). The whole commit results in 30% speed-up compared
to the current master (where threevector divisions are already optimized).

Debug log was axed for two reasons:
 * The collision_time function is conceptually very simple and if
   there is a test to cover it, it does not need debugging messages.
 * The collision_time function is called billions of times and creation
   of logger was 2/3 of its total run time. This is hardly reasonable even in
the debug mode. What if one wants to debug the function?  See point one, simple
test would be much better than those messages.

This refs #2634

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>