parametrization: fix elastic cross sections for kaons
The kaon-nucleon cross sections were using the wrong masses when
calculating plab. This is due to a change to plab_from_s, which assumes
now that both masses are identical if one mass is given. Before it was
assuming that the second mass corresponds to a nucleon. The
parametrization code was updated to work correctly with the new
plab_from_s.

Refs #2745.