Fix a difficulty with anti-strangeness generation
When grand-canonical thermalizer was sampling anti-strangeness,
it could happen that, for example, sampled strangeness is 156 and
initial is 155. If by chance, next sampled particle has strangeness -2,
then the loop condition was fulfilled, but strangeness conservation was broken.
The corrected code will only allow to sample a particle with strangeness -1
in this case.

For baryon number this problem cannot happen, because, it can only be +1,0,-1.

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>