Modify mode sampling to fix isospin symmetry
Looking at particle multiplicities produced by thermalization of
isospin-symmetric system, I found that numbers of pi+- and pi0 are
significantly different. This is because in the modes sampling
(let's take mode 5 as an example) I was doing the following

while (E < total_energy) {
 1. Sample particle
 2. Add its energy to E
 3. Reject it if (strangeness != 0) or (baryon number != 0) or (charge <= 0)
}

The right way is

while (E < total_energy) {
 1. Sample particle with (strangeness != 0) or (baryon number != 0)
 2. Add its energy to E
 3. Reject it if (charge <= 0)
}

Current commit implements the second procedure. The first procedure was
producing not enough charged particles. After that, to conserve energy,
it produced too many neutral particles. So emerged the isospin asymetry.

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>