lgpang:error with InterpolateData on Mac #3806
(1) Mac says 2 components are the same in one vector<double> x, however, I
checked and there are no duplicates vectors. The error maybe caused by
floating a==b treatment in Mac.
(2) The position for this bug was reported to the wrong function. It
should be in the InterpolateDataLinear() function instead of
InterpolateDataSpline() function.

Fix:
(1) Use std::adjacent_find() to make sure there are no duplicates
(2) Remove the (a==b) float equal judgement
(3) Provide the true position where the error message is reported.