add formation time to list Modus #3816
The previous list modus hard set the formation time to start_time_.
This update fixed it with formation time set to the time when particles
get out of the freeze out hypersurface in hydro.

The particles are also anti-freestreamed from their formation time to
start_time_ such that smash can start with the same time.

The cross_section_scaling_factor is set to 0 which means the hadrons
will not scatter with others or decay before formation.