change start_time_ in list Modus #3816
The start_time_ was hard setted in config.yaml for List Modus.
However, it is better to set start_time_ to the earliest formation
time of all the particles. This will save a lot of time by free
streaming the particles from t= earliest_formation_time than t=0 fm.

For particle lists from another SMASH run, there is no
anti-freestreaming, the start_time_ is also changed to the stop time of
the previous simulation.