Improve tabulation limits for rho
This fixes #4148, where omega meson was getting wrong values of rho(m),
because tabulation range was very small. The length of tabulation range
is now chosen as max(2 GeV, 10*maximal width of resonance decay products), while
previously it was 10*maximal width of resonance decay products.

This is still heuristics that can in principle fail. I have checked that
extrapolations still happen all the time, however they always seem to be far right
from the resonance poles, which is ok.

A case breaking the current implementation is a massive resonance with low
threshold and narrow decay products.
Example: M_R = 5 GeV, M_threshold = 2 GeV, one of the products is omega,
another is stable. Then tabulation range will be [2, 4] GeV, where
the upper limit is not even reaching the pole, so the error is guaranteed.

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>