Use a higher-order integration algorithm
This fixes numerical problems that occured when calculating the spectral
integrals of decays into an f₀(1370).

According to the GSL documentation, a higher order algorithm improves
accuracy for smooth functions at the cost of performance for
discontinous functions. I think it is safe to increase the order, since
our functions should usually be smooth, but it might decrease
performance.