Fix det.bal. factor: use real, not pole mass for pcm
In the detailed balance factor for KN-KDelta reaction, pole masses were used
to compute pcm. This does not correspond to det.bal formula and luckily
was causing a crash when the effective Delta mas was below pole mass.

Now the right center-of-mass momentum is computed beforehand
and is passed to det.bal. calculation.

This refs #4086

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>