Propagate_until: fix t_end == t_current case
Because of numerics it was possible that t_current is slightly larger
than t_end. Then the code was crashing, beacuse negative timestep is
not allowed.

This commit fixes the problem by explicitly checking if t_current > t_end.
If the difference is significant then the error message is thrown.

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>