integrate: Set maximum number of subintervals to workspace size
Before it was possible to increase the workspace size, but it did not have any
effect, because the number of maximal subintervals was hardcoded to 500. Now
the workspace size and the number of maximal subintervals is the same. Because
the default workspace size is 1000, this doubles the maximal number of
suberintervals. This does not seem to significantly affect runtime.

The const correctness of Integrator was improved as well.

Refs #4285.