Fix a bug in ignoring interacted nucleons
Before nucleons were never marked as interacted, because it was only
considering invalid actions that never happen by defintion.

This commit fixes that issue by moving the code out of the if statement
so that it runs for *valid* actions. The code was also simplified to use
a for loop that works for any number of incoming particles, as suggested
by Dima.

Fixes and refs #4391