Use better integration algorithm
The new algorithm is also provided by GSL, but it is faster and more
robust. It seems to finally fix the problems with the normalization of
the spectral functions. Also, SMASH is now a bit faster with the default
config (because it takes less time to normalize the spectral functions).

Floating point traps are disabled before calling the integration
routine. This should be fine, because the algorithm is robust enough to
deal with infinties and NaNs.

Fixes and refs #4462.