Fix compiler warning
The variable was initialized in the first iteration of the loop, so the
code is correct. However, the compiler is not clever enough to see this,
so it generates a warning about a possibly uninitialized value. The
easiest way is to initialize it with a value that will be overwritten
anyway.