Finalized formation times according to philosophy
The formation time of the products of a decay is now either the time of execution or, if larger, the formation time of the resonance.
The formation time of product(s) of a scattering is now the time of scattering, or, if larger, the largest formation time of the parent particles.