Implement string fragmentation in GiBUU way
Namely, the scatterings of the low c.m. energies are all through resonances,
while those of the high c.m. energies are all through string fragmentation.
And the scatterings of the medium c.m. energies can be through both the methods
by a probability. In the current version the midium energy region is 4-5 GeV
for NN scattering and 2-2.6 GeV for Npi scattering. The probability for string
fragmentation increases linearly form 0 to 1 in the midium energy region. The total
cross sections at high eneriges are parametrized in the particle data book 2016.
And the parametrization of the cross sections for nucleon pion elastic scatterings
are stolen from GiBUU source code. The program crashes some time for unknown reason
which is not easy to check, since the random seed of pythia is getting involved.

Signed-off-by: Feng Li <fengli@fias.uni-frankfurt.de>