Introduces minimum mass for non zero spectral function
The function return the minimal mass, where the spectral function
is non-zero. This is on the one hand necessary to prevent sampling
a sqrt(s) that only allows mass sampling in region where the
spectral function is zero everywhere and can on the other hand be
used to set more efficient lower boundaries for sampling in
general.

this refs #4965