Fixed formation time check in decay action finder
The formation time was wrongly checked against the acutal
time of the particle, but should be checked against the acutal
time plus the decay time.

This lead to an increase of the lifetime of short-lived particles
like rho and delta and consequently to an increase in the dilepton
yiels.

this refs #4976