Make binary output backwards compatible
The binary header should have the format version followed by the format
variant, because the least significand byte is the first one on little
endian architectures (which are the only ones currently used to run
SMASH). This makes sure that the binary output can still be read by
older analysis scripts.

This also fixes the failing test.