Fix the case of box without interactions
Previously if there were no actionfinders created in the box modus,
SMASH was hanging because the grid was incorrectly initialized.
Max_transvers_distance_sqr was set to a large number, then used
to compute number of cells in the grid, only to obtain 0.

Now max_transvers_distance_sqr is set to a reasonable number
regardless if any actionfinders exist.

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>