improve piN collision
1) Modify the data of the effective piN elastic cross section,
which are from some indirect processes such as resonance formation
and decay, in the data file. It was once obtained through resonance
only, but the contribution from string fragmentation should also
be taken into account.

2) Improve the algorithm for subtracting the cross section of the
indirect elastic scattering from the parametrized elastic cross section.

3) The results suggest that the in the intermediate energy region,
the string fragmentation should trigger faster with energy for pion
nucleon scattering, otherwise, the results are below the experimental
data. So a square root function, instead of a linear function, is used
to trigger the string fragmentation in the intermediate energy.

Signed-off-by: Feng Li <fengli@fias.uni-frankfurt.de>