Comment out all thread_local in SMASH
Thread_local turned to be dangerous due to the known gcc bug
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=60702, which
seems to be fixed in clang, but not in gcc including the newest version.
This bug caused the issue #3075.

Since SMASH is not using parallelization anyway, thread_local statements
were commented out. This allows to get rid of them without removing the
message that they give to a code reader.

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>