Improve Lambda K branching ratios further
When can resolve the tension between the pi N -> Lambda K and the N N ->
Lambda K cross section (which are both scaling with N* -> Lambda K) by
changing the N* -> pi N channel within the freedom given by the PDG and
HADES PWA data. This led to an overestimation of the pi N cross section
due to the Delta(1950) peak. This has been fixed by adding a made-up
decay channel that contributes the branching ratio missing in the
experimental data (before the branching ratios were rescaled to 1,
leading to a pi N branching ratio that was too large).