Improve branching ratios for Sigma production
This used the N N -> Sigma N K and pi N -> Sigma K cross section to
improve the N*, Delta* -> K Sigma, pi N branching ratios.

There is some tension between p pi- -> Sigma- K+ and p pi+ -> Sigma+ K+
affecting the Delta(1930) and Delta(1950) branching ratios, that
unfortunately cannot be reconciled.

The Sigma production in N N might be a bit to high, but it is hard to
tell due to the huge error bars of the data.