Make only baryon be affected by the potentials
For hyperons:
Lambda and Sigma feel 2/3 of the Skyrme force since they carry 2
light (u or d) quarks
Xi feels 1/3 of the Skyrme force.
Omega is not affected by the Skyrme force.
All the hyperons are not affected by the symmetry force.

That's how the hyperons are treated in RVUU.

Signed-off-by: Feng Li <fengli@fias.uni-frankfurt.de>