Remove acts_on from potential_gradient
This refs #5397

This commmit fixes the following bug. There was a difference
between treatment of symmetry potential with and without
lattice - two separate functions were taking care of this and they
were not doing the same thing. Now the logic of how potential
act on specific particle type is at one place - in the propagate.cc
in the function update_momenta. Potential_gradient from potential
class now returns a pair of gradients: for Skyrme and symmetry potentials.

Signed-off-by: Dmytro Oliinychenko <oliiny@fias.uni-frankfurt.de>