Improve strangeness exchange
The parametrizations for K- p -> Lambda pi0 and K- p -> Sigma0 pi0 were
improved by fitting the background to Landolt-Börnstein data. It seems
that the UrQMD parametrization only considered the sum of the two
reactions, not its total. For the other channels, we still use the UrQMD
parametrization because it had good results.

A few resonances related to strangeness exchange have been tweaked
further.